dev v0.2-experimental added ExecutorService to handle queries updated SendV2.java updated Read.java updated Anti-Spam updated User entity: not using float for resource economy updated THandler

+++ src/main/java/com/serezka/jpt/database/model/authorization/User.java
@@ -24,12 +24,10 @@ public class User {
 
     // -> bot settings for user
     Role role = Role.DEFAULT;
-    // <
 
     // -> gpt settings for user
-    Double temperature = 0.5;
+    Float temperature = 0.5F;
     Long chat = 0L;
-    // <
 
     @AllArgsConstructor
     @Getter
